requirements.txt
streamlit
pandas
Pillow
python-dotenv
cryptography
pytest
pytest-mock